+++ src/lib/hue-api/api.ts
@@ -13,6 +13,7 @@ import { delay, normalizeBridgeId } from "../../util.js";
 import LightResource from "./light-resource.js";
 import { AuthenticateResult, AuthenticateSuccess, HubConfig } from "./types.js";
 import GroupResource from "./group-resource.js";
+import SceneResource from "./scene-resource.js";
 
 const MAX_RETRIES = 3;
 
@@ -48,12 +49,14 @@ class HueApi implements ResourceApi {
   private hubUrl?: string;
   public readonly lightResource: LightResource;
   public readonly groupResource: GroupResource;
+  public readonly sceneResource: SceneResource;
   private axiosInstance: AxiosInstance;
 
   constructor(hubUrl?: string, requestTimeout: number = 1500) {
     this.hubUrl = hubUrl;
     this.lightResource = new LightResource(this);
     this.groupResource = new GroupResource(this, this.lightResource);
+    this.sceneResource = new SceneResource(this);
     this.axiosInstance = axios.create({
       baseURL: this.hubUrl,
       timeout: requestTimeout,